pyproject.toml
[project]
name = "haloarchaea-metagenomics"
version = "0.1.0"
description = "A high-throughput Snakemake workflow for Haloarchaea metagenomics analysis."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "snakemake>=8.0.0",
    "pandas",
    "biopython",
    "snakemake-executor-plugin-slurm"
]
